rename client to net
